+++ resources/views/components/icon/align-center.blade.php
@@ -1,0 +1,1 @@
+<svg {{ $attributes }} xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 20 20"><g transform="translate(-2 -2)"><path d="M21,2H3A1,1,0,0,0,3,4H21a1,1,0,0,0,0-2Z"/><path d="M17,10a1,1,0,0,0,0-2H7a1,1,0,0,0,0,2Z"/><path d="M21,14H3a1,1,0,0,0,0,2H21a1,1,0,0,0,0-2Z"/><path d="M17,22a1,1,0,0,0,0-2H7a1,1,0,0,0,0,2Z"/></g></svg>