+++ resources/views/components/input/trix.blade.php
@@ -1,31 +1,24 @@
-@props([ 'for' => null, 'label' => null, 'helpText' => null, 'rowClass' => null, 'inline' => false, 'req' => false, 'initialValue' => null, ])
-
-@php
-    // if wire:model exists even if it is null then isLivewire = true
-    $isLivewire = !empty($attributes->wire('model')->value) ? true : false;
-@endphp
+@props([ 'for' => null, 'label' => null, 'helpText' => null, 'rowClass' => null, 'inline' => false, 'req' => false ])
 
 <div class='frm-row {{ $inline ? 'inline' : '' }} {{ $rowClass }}'
     x-data="{
-        @if($isLivewire)
-            value: @entangle($attributes->wire('model')),
-        @else
-            value: '<?= $initialValue ?>',
-        @endif
-            isFocused() { return document.activeElement !== this.$refs.trix },
-            setValue() { this.$refs.trix.editor.loadHTML(this.value) },
-        }"
-    x-init="setValue(); $watch('value', () => isFocused() && setValue())"
+        value: @entangle($attributes->wire('model')),
+        isFocused() { return document.activeElement !== this.$refs.trix },
+        setValue() { this.$refs.trix.editor.loadHTML(this.value) },
+    }"
+    x-init="setValue();
+        $watch('value', () => isFocused() && setValue())
+        var length = $refs.trix.editor.getDocument().toString().length
+        $refs.trix.editor.setSelectedRange(length - 1)"
     x-on:trix-change="value = $event.target.value"
-    wire:ignore
->
+    wire:ignore>
 
     @isset($label)
         <label @error($for) class="txt-red fw-9 " @enderror for="{{ $for }}">
             {{ Str::title($label) }} @if ($req) <span class='txt-red'>*</span> @endif </label>
     @endisset
 
-    <div  {{ $attributes->class(['w-full'])->whereDoesntStartWith('wire:model') }}>
+    <div {{ $attributes->class(['w-full'])->whereDoesntStartWith('wire:model') }}>
 
         @isset($helpText)
             <div class="help my-05 txt-muted"> <small>{{ $helpText }}</small> </div>
@@ -33,7 +26,7 @@
 
         <input id="{{ $for }}" type="hidden" name="{{ $for }}">
 
-        <trix-editor x-ref="trix" input="{{ $for }}" ></trix-editor>
+        <trix-editor x-ref="trix" input="{{ $for }}"></trix-editor>
 
         @error($for)
             <small class="txt-red" role="alert"> {{ $message }} </small>